+++ test/resource_list_race_test.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:filament_mobile/data/paginated_records.dart';
+import 'package:filament_mobile/data/action_result.dart';
 import 'package:filament_mobile/data/resource_data_source.dart';
 import 'package:filament_mobile/data/resource_record.dart';
 import 'package:filament_mobile/data/write_result.dart';
@@ -67,6 +68,13 @@ class _GatedSource implements ResourceDataSource {
   @override
   Future<WriteResult> destroy(String resourceKey, Object id) =>
       throw UnimplementedError();
+
+  @override
+  Future<ActionResult> runAction(
+    String resourceKey,
+    Object id,
+    String action,
+  ) => throw UnimplementedError();
 
   @override
   Future<OptionsPage> options(